корректировка таблиц

+++ src/main/java/dev/pro/shelter/repository/CatRepository.java
@@ -3,5 +3,5 @@ package dev.pro.shelter.repository;
 import dev.pro.shelter.model.Cat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CatRepository extends JpaRepository<Cat, Long> {
+public interface CatRepository extends JpaRepository<Cat, Integer> {
 }
